Seperated the json data
